worked towards Android input
also made platform enabler (changes what gameobjects are enabled based on what platform it is on). would like to note that onscreen stick and onscreen button have either really weird behaviour or are broken in this project, it would have been way easier to just make a custom joystick. also polished ui a tiny bit and fixed a bug or 2